Safari-609.1.11.1/LayoutTests/accessibility/ios-simulator/abbreviation.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script>
var successfullyParsed = false;
</script>
</head>
<body id="body">

<abbr id="content" title='Internal Revenue Service'>IRS</abbr>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that expanded text values can be retrieved.");

    if (window.accessibilityController) {

        var text = accessibilityController.accessibleElementById("content").childAtIndex(0);
        
        shouldBe("text.description", "'AXLabel: IRS'");
        shouldBe("text.stringAttributeValue('AXExpandedTextValue')", "'Internal Revenue Service'");
    }

    successfullyParsed = true;
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
