LayoutTests/accessibility/ios-simulator/abbreviation-expected.txt

IRS
This tests that expanded text values can be retrieved.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.description is 'AXLabel: IRS'
PASS text.stringAttributeValue('AXExpandedTextValue') is 'Internal Revenue Service'
PASS successfullyParsed is true

TEST COMPLETE
